+++ storage/table/fsm/query_test.go
@@ -31,7 +31,7 @@ func TestFSM_Lookup(t *testing.T) {
 			req: &proto.RequestOp_Range{
 				Key: []byte("Hello"),
 			},
-			wantErr: true,
+			want: &proto.ResponseOp_Range{},
 		},
 		{
 			name: "Lookup full DB with non-existent key",
@@ -41,8 +41,7 @@ func TestFSM_Lookup(t *testing.T) {
 			req: &proto.RequestOp_Range{
 				Key: []byte("Hello"),
 			},
-
-			wantErr: true,
+			want: &proto.ResponseOp_Range{},
 		},
 		{
 			name: "Lookup full DB with existing key",
@@ -52,7 +51,6 @@ func TestFSM_Lookup(t *testing.T) {
 			req: &proto.RequestOp_Range{
 				Key: []byte(fmt.Sprintf(testKeyFormat, 0)),
 			},
-
 			want: &proto.ResponseOp_Range{
 				Kvs: []*proto.KeyValue{
 					{
@@ -153,7 +151,12 @@ func TestFSM_Lookup_Txn(t *testing.T) {
 					wrapRequestOp(&proto.RequestOp_Range{Key: []byte("Hello")}),
 				},
 			},
-			wantErr: true,
+			want: &proto.TxnResponse{
+				Succeeded: true,
+				Responses: []*proto.ResponseOp{
+					wrapResponseOp(&proto.ResponseOp_Range{}),
+				},
+			},
 		},
 		{
 			name: "Lookup full DB with non-existent key",
@@ -165,7 +168,12 @@ func TestFSM_Lookup_Txn(t *testing.T) {
 					wrapRequestOp(&proto.RequestOp_Range{Key: []byte("Hello")}),
 				},
 			},
-			wantErr: true,
+			want: &proto.TxnResponse{
+				Succeeded: true,
+				Responses: []*proto.ResponseOp{
+					wrapResponseOp(&proto.ResponseOp_Range{}),
+				},
+			},
 		},
 		{
 			name: "Lookup full DB with existing key",